Refactor: move socket, replay, and send logic into custom hooks
