+++ src/App.js
@@ -1,13 +1,14 @@
-import { useEffect, useState } from 'react';
-import io from 'socket.io-client';
+import { useState } from 'react';
 import NetworkVisualization from './components/NetworkVisualization';
 import TopologyEditor from './components/TopologyEditor';
 import NodeSelector from './components/NodeSelector';
 import TabbedMessagePanel from './components/TabbedMessagePanel';
 import GraphMetrics from './components/GraphMetrics';
+import useNetworkSocket from './hooks/useNetworkSocket';
+import useSendMessage from './hooks/useSendMessage';
+import useReplaySimulation from './hooks/useReplaySimulation';
 
-const socket = io('http://localhost:8000');
-const MAX_PARALLEL_MESSAGES = 2;
+const COLORS = ['#e91e63', '#2196f3', '#4caf50', '#ff9800', '#9c27b0'];
 
 function App() {
   const [nodes, setNodes] = useState([]);
@@ -24,182 +25,50 @@ function App() {
   const [loading, setLoading] = useState(true);
   const [pathsInFlight, setPathsInFlight] = useState([]);
   const [packetColors, setPacketColors] = useState({});
-  const [logs, setLogs] = useState({});
-  const [paths, setPaths] = useState({});
   const [activeSimId, setActiveSimId] = useState(null);
-  const [metricsBySim, setMetricsBySim] = useState({});
-  const [unreadCounts, setUnreadCounts] = useState({});
-  const [nodeSnapshots, setNodeSnapshots] = useState({});
-  const [replayState, setReplayState] = useState({ simId: null, index: 0 });
 
-  const COLORS = ['#e91e63', '#2196f3', '#4caf50', '#ff9800', '#9c27b0'];
+  const {
+    socket,
+    logs,
+    paths,
+    metricsBySim,
+    unreadCounts,
+    nodeSnapshots,
+  } = useNetworkSocket(activeSimId, setNodes, setEdges, setLoading);
 
-  useEffect(() => {
-    socket.on('networkUpdate', ({ message, nodes, edges, path, colorId, simulationId }) => {
-      if (!simulationId) return;
+  const sendMessage = useSendMessage({
+    socket,
+    graph,
+    sourceNode,
+    destinationNode,
+    pathsInFlight,
+    setPathsInFlight,
+    setPacketColors,
+    COLORS,
+  });
 
-      const tag = colorId != null ? `[${colorId}] ` : '';
-      const fullMessage = tag + message;
-
-      setLogs((prev) => ({
-        ...prev,
-        [simulationId]: [...(prev[simulationId] || []), fullMessage],
-      }));
-
-      if (simulationId !== activeSimId) {
-        setUnreadCounts((prev) => ({
-          ...prev,
-          [simulationId]: (prev[simulationId] || 0) + 1,
-        }));
-      }
-
-      if (Array.isArray(path)) {
-        setPaths((prev) => ({
-          ...prev,
-          [simulationId]: path,
-        }));
-      }
-
-      const safeNodes = Array.isArray(nodes) ? nodes.filter(n => n && n.id) : [];
-      const safeEdges = Array.isArray(edges) ? edges.filter(e => e && e.from && e.to) : [];
-
-      setNodes(safeNodes);
-      setEdges(safeEdges);
-      setLoading(false);
-
-      const retries = message?.includes('(retry') ? 1 : 0;
-      const drops = message?.includes('❌') ? 1 : 0;
-
-      setMetricsBySim((prev) => ({
-        ...prev,
-        [simulationId]: {
-          pathLength: safeNodes.filter(n => n.color === '#4caf50' || n.color === '#f44336').length,
-          totalCost: safeEdges
-            .filter(e => e.color?.color === '#4caf50')
-            .reduce((sum, e) => sum + parseInt(e.label), 0),
-          retries: (prev[simulationId]?.retries || 0) + retries,
-          drops: (prev[simulationId]?.drops || 0) + drops,
-          nodeCount: safeNodes.length,
-          linkCount: safeEdges.length,
-        },
-      }));
-
-      setNodeSnapshots((prev) => ({
-        ...prev,
-        [simulationId]: [...(prev[simulationId] || []), safeNodes],
-      }));
-    });
-
-    socket.on('connect_error', (error) => {
-      console.error('Connection Error:', error);
-      setLoading(false);
-    });
-
-    return () => {
-      socket.off('networkUpdate');
-      socket.off('connect_error');
-    };
-  }, [activeSimId]);
-
-  const sendMessage = () => {
-    if (pathsInFlight.length >= MAX_PARALLEL_MESSAGES) {
-      alert('Please wait – packet limit reached.');
-      return;
-    }
-
-    if (!sourceNode || !destinationNode) {
-      alert('Please select both source and destination.');
-      return;
-    }
-
-    if (!graph[sourceNode] || !graph[destinationNode]) {
-      alert('Selected nodes do not exist in the graph.');
-      return;
-    }
-
-    if (sourceNode === destinationNode) {
-      alert('Source and destination cannot be the same.');
-      return;
-    }
-
-    const newId = pathsInFlight.length;
-    const assignedColor = COLORS[newId % COLORS.length];
-
-    setPathsInFlight((prev) => [...prev, newId]);
-    setPacketColors((prev) => ({ ...prev, [newId]: assignedColor }));
-
-    socket.emit('sendMessage', {
-      from: sourceNode,
-      to: destinationNode,
-      graph,
-      colorId: newId,
-    });
-
-    setTimeout(() => {
-      setPathsInFlight((prev) => prev.filter((id) => id !== newId));
-    }, 6000);
-  };
-
-  const replaySimulation = (simId) => {
-    const snapshots = nodeSnapshots[simId];
-    if (!snapshots) return;
-
-    let i = 0;
-    setReplayState({ simId, index: 0 });
-
-    const interval = setInterval(() => {
-      if (i >= snapshots.length) return clearInterval(interval);
-      setNodes(snapshots[i]);
-
-      // Set the currently highlighted node
-      const activeNode = snapshots[i].find(n => n.color === '#f44336');
-      if (activeNode) {
-        const updated = snapshots[i].map(n =>
-          n.id === activeNode.id
-            ? { ...n, color: '#ffeb3b' } // yellow highlight
-            : n
-        );
-        setNodes(updated);
-      }
-
-      setReplayState({ simId, index: i });
-      i++;
-    }, 800);
-  };
+  const { replayState, replaySimulation } = useReplaySimulation(nodeSnapshots, setNodes);
 
   const onCloseTab = (simId) => {
-    setLogs((prev) => {
-      const updated = { ...prev };
-      delete updated[simId];
-      return updated;
-    });
+    // clean up per-simulation data
+    ['logs', 'paths', 'metricsBySim', 'unreadCounts', 'nodeSnapshots'].forEach((key) => {
+      const setter = {
+        logs: () => socket.setLogs,
+        paths: () => socket.setPaths,
+        metricsBySim: () => socket.setMetricsBySim,
+        unreadCounts: () => socket.setUnreadCounts,
+        nodeSnapshots: () => socket.setNodeSnapshots,
+      }[key];
 
-    setPaths((prev) => {
-      const updated = { ...prev };
-      delete updated[simId];
-      return updated;
-    });
-
-    setMetricsBySim((prev) => {
-      const updated = { ...prev };
-      delete updated[simId];
-      return updated;
-    });
-
-    setUnreadCounts((prev) => {
-      const updated = { ...prev };
-      delete updated[simId];
-      return updated;
-    });
-
-    setNodeSnapshots((prev) => {
-      const updated = { ...prev };
-      delete updated[simId];
-      return updated;
+      setter((prev) => {
+        const updated = { ...prev };
+        delete updated[simId];
+        return updated;
+      });
     });
 
     if (simId === activeSimId) {
-      const remaining = Object.keys(logs).filter(id => id !== simId);
+      const remaining = Object.keys(logs).filter((id) => id !== simId);
       setActiveSimId(remaining[0] || null);
     }
   };
@@ -228,43 +97,37 @@ function App() {
       />
 
       <div style={{ textAlign: 'center', marginBottom: '20px' }}>
-        <button onClick={sendMessage} disabled={pathsInFlight.length >= MAX_PARALLEL_MESSAGES}>
-          {pathsInFlight.length >= MAX_PARALLEL_MESSAGES ? '⏳ Sending...' : '📤 Send Message'}
+        <button onClick={sendMessage} disabled={pathsInFlight.length >= 2}>
+          {pathsInFlight.length >= 2 ? '⏳ Sending...' : '📤 Send Message'}
         </button>
       </div>
 
       {loading ? (
         <p>Loading...</p>
       ) : (
-        <>
-          <div style={{ display: 'flex', alignItems: 'flex-start', gap: '30px' }}>
-            <div style={{ flex: 1 }}>
-              <NetworkVisualization
-                nodes={nodes}
-                edges={edges}
-                animatePath={Object.values(paths)}
-              />
-              {replayState.simId && (
-                <div style={{ marginTop: '10px', textAlign: 'center' }}>
-                  🕒 Step {replayState.index + 1} / {nodeSnapshots[replayState.simId]?.length || 0}
-                </div>
-              )}
-            </div>
-            <div style={{ width: '450px' }}>
-              <TabbedMessagePanel
-                logs={logs}
-                packetColors={packetColors}
-                setActiveSimId={setActiveSimId}
-                activeSimId={activeSimId}
-                unreadCounts={unreadCounts}
-                setUnreadCounts={setUnreadCounts}
-                onCloseTab={onCloseTab}
-                replaySimulation={replaySimulation}
-              />
-              <GraphMetrics metrics={currentMetrics} activeSimId={activeSimId} />
-            </div>
+        <div style={{ display: 'flex', alignItems: 'flex-start', gap: '30px' }}>
+          <div style={{ flex: 1 }}>
+            <NetworkVisualization nodes={nodes} edges={edges} animatePath={Object.values(paths)} />
+            {replayState.simId && (
+              <div style={{ marginTop: '10px', textAlign: 'center' }}>
+                🕒 Step {replayState.index + 1} / {nodeSnapshots[replayState.simId]?.length || 0}
+              </div>
+            )}
           </div>
-        </>
+          <div style={{ width: '450px' }}>
+            <TabbedMessagePanel
+              logs={logs}
+              packetColors={packetColors}
+              setActiveSimId={setActiveSimId}
+              activeSimId={activeSimId}
+              unreadCounts={unreadCounts}
+              setUnreadCounts={() => {}}
+              onCloseTab={onCloseTab}
+              replaySimulation={replaySimulation}
+            />
+            <GraphMetrics metrics={currentMetrics} activeSimId={activeSimId} />
+          </div>
+        </div>
       )}
     </div>
   );
